feat(migrations): add native_query hypertable and float interval support

+++ api/app/databases/configs/models.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel, ConfigDict
 class ConfigItem(BaseModel):
     id: int
     name: str
-    interval: int
+    interval: float
     is_paused: bool
 
     model_config = ConfigDict(from_attributes=True)
@@ -14,4 +14,4 @@ class ConfigItem(BaseModel):
 
 class ConfigPatch(BaseModel):
     is_paused: Optional[bool] = None
-    interval: Optional[int] = None
+    interval: Optional[float] = None
